pyproject.toml
[project]
name = "enochecker3"
version = "0.13.4"
authors = [
	{ name = "Louis Burda", email = "dev@sinitax.com" },
	{ name = "Lucas Druschke", email = "ldruschk@posteo.de" }
]
description = "Library for building EnoEngine-compatible async checkers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"fastapi>=0.128.0,<0.129.0",
	"gunicorn==24.1.1",
	"httpx>=0.28.0,<0.29.0",
	"pymongo>=4.0.0,<5.0.0",
	"opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
	"opentelemetry-instrumentation-httpx>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
	"uvicorn>=0.40.0,<0.41.0",
	"enochecker-core>=0.13.0,<0.14.0"
]

[dependency-groups]
dev = [
	"ruff>=0.12.10",
	"coverage>=7.10.5",
	"pytest>=8.4.1",
	"pytest-asyncio>=1.1.0",
	"mypy>=1.17.1",
]

[tool.ruff.lint]
ignore = ["E731", "E722"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
exclude = []

[tool.setuptools.package-data]
"enochecker3" = ["py.typed"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-s -rP"
testpaths = [
	"tests"
]

